+++ app/blogs/[slug]/page.tsx
@@ -3,9 +3,9 @@ import Image from 'next/image';
 import { Box, Flex, Title } from '@mantine/core';
 import fs from 'fs';
 import path from 'path';
-import Comments from '@/components/blogs/Comments';
-import { MdxComponents } from '@/components/blogs/mdxComponents/MdxComponents';
-import MDXContent from '@/components/blogs/mdxContent/MDXContent';
+import Comments from '@/components/comments/Comments';
+import { MdxComponents } from '@/components/mdxComponents/MdxComponents';
+import MDXContent from '@/components/mdxContent/MDXContent';
 import { BLOG_CONTENTS_DIR } from '@/constants/blogPosts';
 import getBlogPost from '@/utils/getBlogPost';
 import styles from './page.module.scss';
